test(dolt): fix CLI-route logging test's self-skip construction defect (refs be-9i0yq.2, review be-j7eon)
TestPushPullLogCLIRouteForGitProtocolRemote used plain New(), which in this
harness resolves to a client of the shared containerized test server
(testmain_test.go) rather than a store with its own local .dolt directory.
CLIDir() and hasCLIDatabase() were therefore always empty/false, so the test
always hit its own skip branch and never exercised CLI routing at all.

Move it into a new integration-tagged file and give it its own local
dolt sql-server via startLocalDoltServer, the same arrangement
TestGitRemoteEmbeddedPushPull uses, so the store has a real on-disk
CLI-accessible database. Confirmed the test now runs (not skips) and
passes, asserting the CLI-route log lines it always meant to check.

TestPushPullLogSQLRouteForFileRemote is untouched and keeps its original
build classification (no integration tag) since it works correctly as-is
against the shared test server.
